Update accounting application

+++ routes/hr.php
@@ -11,20 +11,20 @@ use App\Http\Controllers\SalaryController;
 Route::middleware(['auth', 'verified', 'company'])
     ->group(function (): void {
 
-        Route::get('/employees', [EmployeeController::class, 'index'])->name('employees.index');
-        Route::get('/employees/create', [EmployeeController::class, 'create'])->name('employees.create');
-        Route::post('/employees', [EmployeeController::class, 'store'])->name('employees.store');
-        Route::get('/employees/{employee}', [EmployeeController::class, 'show'])->name('employees.show');
-        Route::get('/employees/{employee}/edit', [EmployeeController::class, 'edit'])->name('employees.edit');
-        Route::put('/employees/{employee}', [EmployeeController::class, 'update'])->name('employees.update');
-        Route::patch('/employees/{employee}', [EmployeeController::class, 'update']);
-        Route::delete('/employees/{employee}', [EmployeeController::class, 'destroy'])->name('employees.destroy');
+        Route::get('/employees', [EmployeeController::class, 'index'])->middleware('can-permission:employees.view')->name('employees.index');
+        Route::get('/employees/create', [EmployeeController::class, 'create'])->middleware('can-permission:employees.create')->name('employees.create');
+        Route::post('/employees', [EmployeeController::class, 'store'])->middleware('can-permission:employees.create')->name('employees.store');
+        Route::get('/employees/{employee}', [EmployeeController::class, 'show'])->middleware('can-permission:employees.view')->name('employees.show');
+        Route::get('/employees/{employee}/edit', [EmployeeController::class, 'edit'])->middleware('can-permission:employees.edit')->name('employees.edit');
+        Route::put('/employees/{employee}', [EmployeeController::class, 'update'])->middleware('can-permission:employees.edit')->name('employees.update');
+        Route::patch('/employees/{employee}', [EmployeeController::class, 'update'])->middleware('can-permission:employees.edit');
+        Route::delete('/employees/{employee}', [EmployeeController::class, 'destroy'])->middleware('can-permission:employees.delete')->name('employees.destroy');
 
-        Route::get('/salaries', [SalaryController::class, 'index'])->name('salaries.index');
-        Route::get('/salaries/create', [SalaryController::class, 'create'])->name('salaries.create');
-        Route::post('/salaries', [SalaryController::class, 'store'])->name('salaries.store');
-        Route::get('/salaries/{salary}', [SalaryController::class, 'show'])->name('salaries.show');
-        Route::delete('/salaries/{salary}', [SalaryController::class, 'destroy'])->name('salaries.destroy');
+        Route::get('/salaries', [SalaryController::class, 'index'])->middleware('can-permission:salaries.view')->name('salaries.index');
+        Route::get('/salaries/create', [SalaryController::class, 'create'])->middleware('can-permission:salaries.create')->name('salaries.create');
+        Route::post('/salaries', [SalaryController::class, 'store'])->middleware('can-permission:salaries.create')->name('salaries.store');
+        Route::get('/salaries/{salary}', [SalaryController::class, 'show'])->middleware('can-permission:salaries.view')->name('salaries.show');
+        Route::delete('/salaries/{salary}', [SalaryController::class, 'destroy'])->middleware('can-permission:salaries.delete')->name('salaries.destroy');
         Route::post('/salaries/{salary}/mark-paid', [SalaryController::class, 'markPaid'])
-            ->name('salaries.mark-paid');
+            ->middleware('can-permission:salaries.edit')->name('salaries.mark-paid');
     });
